web-app/requirements.txt
junos_eznc==2.6.7
matplotlib==3.7.3
netgraph==4.12.11
networkx==3.1
numpy==1.25.2
pandas==2.1.1
pyvis==0.3.2
streamlit==1.30.0
streamlit_option_menu==0.3.6
streamlit-ace==0.1.1
nt-lib==0.1.4
PyYAML==6.0.1
PyGithub==2.1.1
requests==2.27.1
yamllint==1.33.0
rarfile==4.1
